internal: move the mappers to their own separate file
